fix(batch-cp): active mask param is uint8_t* (matches the scheduler array)
gcc-15 on the EPYC box promotes incompatible-pointer-types to an error (gcc-11
on the M1 only warned, and the warning slipped through the build grep). The
scheduler's active[] is uint8_t*; align qwen_batch_cp_predict + the internal
layer/step helpers. Behavior unchanged (the mask is only tested for truth).

+++ qwen_tts_batch.h
@@ -62,7 +62,7 @@ int qwen_cp_predict(qwen_tts_ctx_t *ctx, float *talker_hidden, int code0, int *o
  * x86 follow-up, see plan_v4). */
 int qwen_batch_cp_predict(qwen_tts_ctx_t *ctx, qwen_batch_t *bb,
                           const float *talker_hidden, const int *code0, int *out_codes,
-                          const int *active);
+                          const uint8_t *active);
 
 /* Allocate batched buffers + B KV caches from ctx config. kv_max = max frames per
  * chunk. Returns NULL on OOM or if the model isn't bf16 (v1 limitation). */
